seperate user route and some refactoring
