+++ models/index.js
@@ -1,0 +1,9 @@
+export { default as User } from "../models/user.js";
+export { default as Privacy } from "../models/privacy.js";
+export { default as Help } from "../models/help.js";
+export { default as Feedback } from "../models/feedback.js";
+export { default as ReportUser } from "../models/reportUser.js";
+export { default as Home } from "../models/home.js";
+export { default as Review } from "../models/review.js";
+export { default as Reserve } from "../models/reserve.js";
+export { default as General } from "../models/general.js";
