+++ server.js
@@ -1,6 +1,7 @@
 import express from "express";
 import bodyParser from "body-parser";
-import apiRouter from "./routes/api.js";
+import apiRoute from "./routes/api.js";
+import userRoute from "./routes/user.js";
 import { PORT } from "./config/environments.js";
 import "./config/db.js";
 
@@ -8,12 +9,18 @@ const app = express();
 
 app.use("/uploads", express.static("uploads")); // access permission
 app.use(bodyParser.json());
-app.use("/api", apiRouter);
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use("/users", (req, res, next) => {
+  console.log("user route");
+  res.status(200).send("user route");
+  next();
+});
+app.use("/", apiRoute);
 
 app.use((err, req, res, next) => {
   res.status(422).send({ error: err._message });
 });
 
 app.listen(PORT, () => {
-  console.log(`SERVER IS READY ON PORT ${PORT}!`);
+  console.log(`==========> SERVER IS READY ON PORT ${PORT}!`);
 });
